% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okBATHTUB-package.R
\name{assert_okBATHTUB}
\alias{assert_okBATHTUB}
\title{Validate that an object is an okBATHTUB result}
\usage{
assert_okBATHTUB(x, required_step = NULL)
}
\arguments{
\item{x}{Object to check.}

\item{required_step}{If supplied, also checks that the pipeline has
progressed at least to the step preceding \code{required_step}.}
}
\description{
Validate that an object is an okBATHTUB result
}
\keyword{internal}
