% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okBATHTUB-package.R
\name{.assert_nonneg}
\alias{.assert_nonneg}
\title{Assert single non-negative finite numeric}
\usage{
.assert_nonneg(x, name)
}
\description{
Assert single non-negative finite numeric
}
\keyword{internal}
