% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okBATHTUB-package.R
\name{.assert_positive}
\alias{.assert_positive}
\title{Assert single positive finite numeric}
\usage{
.assert_positive(x, name)
}
\description{
Assert single positive finite numeric
}
\keyword{internal}
