% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/awqms.R
\name{.awqms_default_config}
\alias{.awqms_default_config}
\title{Default AWQMS connection configuration}
\usage{
.awqms_default_config()
}
\description{
Default AWQMS connection configuration
}
\keyword{internal}
