% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsi.R
\name{.classify_trophic_state}
\alias{.classify_trophic_state}
\title{Classify trophic state from mean TSI}
\usage{
.classify_trophic_state(tsi)
}
\arguments{
\item{tsi}{Numeric TSI value.}
}
\description{
Classify trophic state from mean TSI
}
\keyword{internal}
