% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okBATHTUB-package.R
\name{print.okBATHTUB}
\alias{print.okBATHTUB}
\title{Print method for okBATHTUB objects}
\usage{
\method{print}{okBATHTUB}(x, ...)
}
\arguments{
\item{x}{An \code{okBATHTUB} object.}

\item{...}{Ignored.}
}
\description{
Print method for okBATHTUB objects
}
