% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okBATHTUB-package.R
\name{summary.okBATHTUB}
\alias{summary.okBATHTUB}
\title{Summary method for okBATHTUB objects}
\usage{
\method{summary}{okBATHTUB}(object, ...)
}
\arguments{
\item{object}{An \code{okBATHTUB} object.}

\item{...}{Ignored.}
}
\description{
Prints a formatted summary of all water quality predictions accumulated
through the pipeline. Most informative after \code{ok_tsi()} has been run.
}
