% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okBATHTUB-package.R
\name{.walker_coefficients}
\alias{.walker_coefficients}
\title{Walker (1996) default coefficient set}
\usage{
.walker_coefficients()
}
\value{
Named list of coefficients.
}
\description{
Nationally-derived empirical coefficients from Walker's BATHTUB model.
These are the default for all okBATHTUB model functions.
}
\keyword{internal}
